backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
websockets==12.0
pydantic==2.6.1
openai==1.12.0
python-magic==0.4.27
Pillow==10.2.0
requests==2.31.0 